feat: 独立图表代理lib，修改示例脚本，移除冗余代码
